oopディレクトリを作成して既存のコードを4つのファイルに分散
